Fix reference to file that doesn't exist.
While reading a page on the web site, I noticed a reference to
"SubmittingPatches", which as far as I can tell was just an outdated
reference to the current CONTRIBUTING doc.  Fix this, make it a link,
and add links for other referenced docs in this file while we're at
it.

Signed-off-by: Russell Bryant <rbryant@redhat.com>