_layouts: Include CSS inline
For some reason, GitHub Pages doesn't seem to be including the assets
like it should be and is replacing them with normalize. I don't know why
this is, but it could be a while before I figure out what's happening.

In the interim, hack around the problem by just moving the CSS inline.
This involves copying 'style.css' to the the '_includes' folder so we
can "include" it inside a '<style>' tag.

Signed-off-by: Stephen Finucane <stephen@that.guru>